revise PCI support
- move PCI support partly to the hardrware indenpendent part
- using the crate `pci_types` to simplify the driver implementations
- revise MMIO driver to support the same interface like the PCI drivers
- remove hardware dependent code within the MMIO interface
- add draft of mmio support for aarch64
- remove clippy warnings
- remove code duplications
- use type InterruptLine to specify the interrupt number

+++ src/arch/x86_64/kernel/core_local.rs
@@ -76,6 +76,6 @@ pub fn set_core_scheduler(scheduler: *mut PerCoreScheduler) {
 	CoreLocal::get().scheduler.set(scheduler);
 }
 
-pub fn increment_irq_counter(irq_no: usize) {
+pub fn increment_irq_counter(irq_no: u8) {
 	CoreLocal::get().irq_statistics.inc(irq_no);
 }
